👌 IMPROVE: Rename Response class filename
