+++ app/loader.php
@@ -2,5 +2,5 @@
 
 require_once 'config.php';
 require_once 'lib/helpers.php';
-require_once 'lib/Response.class.php';
+require_once 'lib/Response.php';
 require_once 'app.php';
